tracer: use internal asset
We use go-bindata to put the object file in a Go file. The generated Go
file is part of the tracer package (pkg/tracer/tcptracer-ebpf.go) and is
included in the git repository, so external projects like Scope can just
vendor the tracer package without worrying about fetching the object
file.

Running 'make' should automatically update the Go file.

CircleCI should error out if the checked-in Go file is stale.