Add (very short) timeout to Dial with IPv6
Connecting to a random IPv6 address will most probably not succeed,
but will timeout eventually.

The default timeout is too high and we don't really care about the
connection succeeding or not, we just need to trigger the kprobe.

This adds a timeout of 1 millisecond to the Dial function for IPv6 so
each connection doesn't hang for a long time.

We didn't notice this before because we haven't tried the guessing
program in an IPv6 network.