offsetguess: accept connections
The server was only listening and not accepting, so we were relying on
the incoming buffer to fit all the connections.

This might cause connection timeouts when the buffer fills, so we now
accept the connections and close them as soon as possible.

To avoid possible races where the connecting side is much faster than
the accepting side and can potentially fill the buffer too, we signal
from the main thread when we are about to make a connection to make sure
the server is ready to accept.

Also, we simplified the server code by moving the listening part out of
the goroutine.