guess: tie the state machine to a single thread
By default, Golang does not guarantee that a go routine will always be
executed on the same system thread (tid). During the guessing work, the
tcp connections could be generated from various threads. The ebpf
program was accepting any connections as long as it come from the
correct process, regardless of threads.

This patch changes this to have stricter rules:
- the ebpf program only accept guessing connections that comes from the
expected pid & tid.
- the userland guessing work takes care of using the thread specified in
the status map. In Golang, this is done with runtime.{Lock,Unlock}OSThread.