offsetguess: workaround missing kretprobes
If the kretprobe for tcp_v4_connect() is configured with a too-low
maxactive, some kretprobe might be missing. In this case, we detect it
and try again.

This is more likely to happen on a single-core VM with a non-preemptive
kernel (CONFIG_PREEMPT not set) because maxactive would be 1 in that
case.

See https://github.com/weaveworks/tcptracer-bpf/issues/24

Based on work from @iaguis.