offsetguess: increase threshold
We've found that Google COS images in GCE were failing when trying to
guess netns. This was because the field is beyond the threshold (one
byte away, actually).

Increase the threshold to 400 to have some breathing room in case more
fields are added to the structs we guess. Offset guessing is done only
once at the beggining and is quite fast, so there's no appreciable loss
of performance.

See
https://github.com/weaveworks/scope/issues/2796#issuecomment-360759884
for an longer explanation.