Add an offset to all timestamps
The timestamps we return are compared against the time that poll()
returned, in github.com/iovisor/gobpf/elf/perf.go.  By adding an
offset we cause iovisor to hold back some events until a later poll,
which means they are less likely to be reported out of order.

The default offset is 0.1ms, which empirically seems to work well for
monitoring consumed by humans.  If this doesn't work for your case,
the variable is exported.